+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added MSP432E4 target (MSP432E401Y and MSP432E411Y). (#1139)
 
 
+### Changed
+
+- probe-rs-cli: more descriptive error messages for ambigous chips
+
 ### Fixed
 
 - probe-rs-cli: fixed `--base-address` having no effect